feat: 회사명을 입력하면 pdf 업로드하는 기능 구현

+++ src/main/java/primer/hackerton/service/S3Service.java
@@ -1,7 +1,6 @@
 package primer.hackerton.service;
 
 import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
@@ -11,12 +10,11 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.retry.annotation.Backoff;
 import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
-import primer.hackerton.domain.report.dto.ReportInfo;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.HttpURLConnection;
 import java.net.URL;
-import java.util.List;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -28,22 +26,41 @@ public class S3Service {
     private final AmazonS3 amazonS3Client;
 
     @Retryable(retryFor = IOException.class, backoff = @Backoff(delay = 1000))
-    public void uploadReportsToS3(String name, String fileUrl, String date) throws IOException, InterruptedException {
+    public String uploadReportsToS3(String name, String fileUrl, String date) throws IOException, InterruptedException {
 
         String key = date + "/" + name + ".pdf";
 
-        try (InputStream inputStream = new URL(fileUrl).openStream()) {
-            ObjectMetadata metadata = new ObjectMetadata();
-            metadata.setContentType("application/pdf");
+        if (amazonS3Client.doesObjectExist(bucketName, key)) {
+            return amazonS3Client.getUrl(bucketName, key).toString();
+        }
 
-            amazonS3Client.putObject(new PutObjectRequest(
-                    bucketName,
-                    key,
-                    inputStream,
-                    metadata
-            ).withCannedAcl(CannedAccessControlList.Private));
+        HttpURLConnection connection = null;
+        try {
+            URL url = new URL(fileUrl);
+            connection = (HttpURLConnection) url.openConnection();
+            int contentLength = connection.getContentLength();
+
+            try (InputStream inputStream = connection.getInputStream()) {
+                ObjectMetadata metadata = new ObjectMetadata();
+                metadata.setContentType("application/pdf");
+                metadata.setContentLength(contentLength);  // 파일 크기 설정
+
+                amazonS3Client.putObject(new PutObjectRequest(
+                        bucketName,
+                        key,
+                        inputStream,
+                        metadata
+                ).withCannedAcl(CannedAccessControlList.Private));
+
+                return amazonS3Client.getUrl(bucketName, key).toString();
+            }
         } catch (IOException e) {
-            log.info(fileUrl + "에러");
+            log.info(fileUrl + " 에러");
+            throw new RuntimeException(e);
+        } finally {
+            if (connection != null) {
+                connection.disconnect();
+            }
         }
     }
 }
